% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFE.R
\name{LFE}
\alias{LFE}
\title{LFE}
\usage{
LFE(xx, yy, T = 5)
}
\arguments{
\item{xx}{model matrix of explanatory variables}

\item{yy}{label vector}

\item{T}{number of instance used to update weights, default to be 5}
}
\value{
\item{w}{new weight matrix after LFE algorithm}
}
\description{
This function performs LFE(Local Feature Extraction) algorithm.
}
\examples{
data(park)
xx<-park$xx
yy<-park$yy
re<-LFE(xx,yy)
print(re)
}
\references{
Sun Y, Wu D. A relief based feature extraction algorithm[C]//Proceedings of the 2008 SIAM International Conference on Data Mining. Society for Industrial and Applied Mathematics, 2008: 188-195.
}
\keyword{LFE}
