% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/park.R
\docType{data}
\name{park}
\alias{park}
\title{Parkinsons Dataset}
\format{
An object of class
}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/parkinsons}{parkinsons}
}
\usage{
data(park)
}
\description{
Parkinsons Dataset
}
\examples{
data(park)
xx <- park$xx
yy <- park$yy

}
\references{
Frank, A. and A. Asuncion. UCI Machine Learning Repository. 2010.
}
\keyword{datasets}
