% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.LFE.R
\name{pred.LFE}
\alias{pred.LFE}
\title{pred.LFE}
\usage{
pred.LFE(w, xx, yy, newx)
}
\arguments{
\item{w}{weight}

\item{xx}{model matrix of explanatory variables}

\item{yy}{label vector}

\item{newx}{model matrix to be preded}
}
\value{
preded labels
}
\description{
This function performs predion for LFE(Local Feature Extraction) algorithm.
}
\examples{
data(park)
xx<-park$xx
yy<-park$yy
w<-LFE(xx,yy)
pred<-pred.LFE(w,xx,yy,xx)
print(pred)
}
\keyword{LFE}
