% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simba.R
\name{Simba}
\alias{Simba}
\title{Simba}
\usage{
Simba(train_xx, train_yy, T = 5)
}
\arguments{
\item{train_xx}{model matrix of explanatory variables}

\item{train_yy}{label vector}

\item{T}{number of instance used to update weights, default to be 5}
}
\value{
\item{w}{new weight after Simba algorithm}
}
\description{
This function performs Simba(Iterative Search Margin Based Algorithm).
}
\examples{
data(park)
xx<-park$xx
yy<-park$yy
re<-Simba(xx,yy)
print(re)
}
\references{
Gilad-Bachrach R, Navot A, Tishby N. Margin based feature selection-theory and algorithms[C]//Proceedings of the twenty-first international conference on Machine learning. ACM, 2004: 43.
}
\keyword{Simba}
