% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.IM4E.R
\name{one.IM4E}
\alias{one.IM4E}
\title{one.IM4E}
\usage{
one.IM4E(train_xx, train_yy, w, sig = 1, lambda = 1)
}
\arguments{
\item{train_xx}{model matrix of explanatory variables}

\item{train_yy}{label vector}

\item{w}{initial weight}

\item{sig}{sigma used in algorithm, default to be 1}

\item{lambda}{lambda used in algorithm, default to be 1}
}
\value{
\item{w}{new weight vector after one loop}

\item{C}{cost after one loop}
}
\description{
This function performs (IM4E)Iterative Margin-Maximization under Max-Min Entropy algorithm for one loop.
}
\keyword{IM4E}
